Move updater.py to _internal
