+++ certbot/_internal/main.py
@@ -29,7 +29,7 @@ from certbot import log
 from certbot._internal import renewal
 from certbot._internal import reporter
 from certbot._internal import storage
-from certbot import updater
+from certbot._internal import updater
 from certbot import util
 from certbot.compat import filesystem
 from certbot.compat import misc
